feat(fp-316): refactor icon structure and scripts / add new icons
