+++ src/lib/components/02-components/banners/flyout/flyout-banner.tsx
@@ -1,8 +1,6 @@
 import { ReactNode, useEffect } from 'react';
 import { cn } from '../../../../utils/tailwind-merge';
-import ArrowRightIcon from '../../../00-tokens/icons/arrow-right-icon';
-import ChevronLeftIcon from '../../../00-tokens/icons/chevron-left-icon';
-import CloseFilledIcon from '../../../00-tokens/icons/close-filled-icon';
+import { ArrowRightIcon, ChevronLeftIcon, CloseFilledIcon } from '../../../00-tokens/icons';
 import Link from '../../../01-elements/link/link';
 import Typography from '../../../01-elements/typography/typography';
 import { BannerProps } from '../masthead/masthead-banner';
