+++ src/lib/components/02-components/menus/menu-links/main-nav-link/main-nav-link.tsx
@@ -1,6 +1,6 @@
 import { ElementType } from 'react';
 import { cn } from '../../../../../utils/tailwind-merge';
-import ArrowBottomIcon from '../../../../00-tokens/icons/arrow-bottom-icon';
+import { ArrowBottomIcon } from '../../../../00-tokens/icons';
 import Link from '../../../../01-elements/link/link';
 
 export type MainNavLinkProps = {
